requirements.txt
tensorflow-gpu==1.15
sklearn
openpyxl
xlutils
matplotlib
xlrd
xlwt